Refactor to separate library scanning from repository access & reduce number of realm instances
