+++ app/src/main/java/net/sigmabeta/chipbox/ui/artist/ArtistListPresenter.kt
@@ -4,13 +4,12 @@ import android.os.Bundle
 import net.sigmabeta.chipbox.R
 import net.sigmabeta.chipbox.dagger.scope.ActivityScoped
 import net.sigmabeta.chipbox.model.domain.Artist
-import net.sigmabeta.chipbox.model.repository.Repository
 import net.sigmabeta.chipbox.ui.BaseView
 import net.sigmabeta.chipbox.ui.FragmentPresenter
 import javax.inject.Inject
 
 @ActivityScoped
-class ArtistListPresenter @Inject constructor(val repository: Repository) : FragmentPresenter() {
+class ArtistListPresenter @Inject constructor() : FragmentPresenter() {
     var view: ArtistListView? = null
 
     var artists: List<Artist>? = null
