+++ app/src/main/java/net/sigmabeta/chipbox/ui/main/MainPresenter.kt
@@ -9,7 +9,6 @@ import net.sigmabeta.chipbox.model.events.GameEvent
 import net.sigmabeta.chipbox.model.events.PositionEvent
 import net.sigmabeta.chipbox.model.events.StateEvent
 import net.sigmabeta.chipbox.model.events.TrackEvent
-import net.sigmabeta.chipbox.model.repository.Repository
 import net.sigmabeta.chipbox.ui.ActivityPresenter
 import net.sigmabeta.chipbox.ui.BaseView
 import net.sigmabeta.chipbox.util.logError
@@ -19,7 +18,7 @@ import javax.inject.Inject
 import javax.inject.Singleton
 
 @Singleton
-class MainPresenter @Inject constructor(val player: Player, val repository: Repository) : ActivityPresenter() {
+class MainPresenter @Inject constructor(val player: Player) : ActivityPresenter() {
     var view: MainView? = null
 
     var state = player.state
