+++ app/src/main/java/net/sigmabeta/chipbox/ui/navigation/NavigationPresenter.kt
@@ -9,7 +9,6 @@ import net.sigmabeta.chipbox.model.events.GameEvent
 import net.sigmabeta.chipbox.model.events.PositionEvent
 import net.sigmabeta.chipbox.model.events.StateEvent
 import net.sigmabeta.chipbox.model.events.TrackEvent
-import net.sigmabeta.chipbox.model.repository.Repository
 import net.sigmabeta.chipbox.ui.ActivityPresenter
 import net.sigmabeta.chipbox.ui.BaseView
 import net.sigmabeta.chipbox.util.logError
@@ -19,7 +18,7 @@ import javax.inject.Inject
 import javax.inject.Singleton
 
 @Singleton
-class NavigationPresenter @Inject constructor(val player: Player, val repository: Repository) : ActivityPresenter() {
+class NavigationPresenter @Inject constructor(val player: Player) : ActivityPresenter() {
     var view: NavigationView? = null
 
     // A property is kept in order to be able to track changes in state.
