+++ app/src/main/java/net/sigmabeta/chipbox/ui/scan/ScanPresenter.kt
@@ -3,7 +3,7 @@ package net.sigmabeta.chipbox.ui.scan
 import android.os.Bundle
 import android.util.Log
 import net.sigmabeta.chipbox.model.events.FileScanEvent
-import net.sigmabeta.chipbox.model.repository.Repository
+import net.sigmabeta.chipbox.model.repository.LibraryScanner
 import net.sigmabeta.chipbox.ui.ActivityPresenter
 import net.sigmabeta.chipbox.ui.BaseView
 import net.sigmabeta.chipbox.util.logError
@@ -15,7 +15,7 @@ import javax.inject.Singleton
 
 
 @Singleton
-class ScanPresenter @Inject constructor(val repository: Repository) : ActivityPresenter() {
+class ScanPresenter @Inject constructor(val scanner: LibraryScanner) : ActivityPresenter() {
     var view: ScanView? = null
 
     var filesAdded = 0
@@ -34,7 +34,7 @@ class ScanPresenter @Inject constructor(val repository: Repository) : ActivityPr
     override fun onTempDestroy() = Unit
 
     override fun setup(arguments: Bundle?) {
-        repository.scanLibrary()
+        scanner.scanLibrary()
                 .buffer(17, TimeUnit.MILLISECONDS)
                 .subscribeOn(Schedulers.io())
                 .observeOn(AndroidSchedulers.mainThread())
